% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet-name.R
\name{max_pribreed}
\alias{max_pribreed}
\title{The preference of pet owners in Seattle.}
\usage{
max_pribreed(pri_breed)
}
\arguments{
\item{The}{\strong{pri_breed} is the primary breed of the particular species.}
}
\value{
The function will return a 2 columns data frame that show the zipcode and number of
particular primary breed.
}
\description{
Enter a primary breed to see which 5 zip code districts have the highest number of
that particular primary breed.
}
