% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet-name.R
\name{name_of_the_year}
\alias{name_of_the_year}
\title{The top 5 pet name of a particular year.}
\usage{
name_of_the_year(pet_name, top_year)
}
\arguments{
\item{pet_name}{is the name you want to check.}

\item{top_year}{is the year you want to to check.}
}
\value{
The function will return a 2 columns data frame that show the top 5 pet names and the
frequency of the name of that year, if the input name is not a in the top 5 list,
a sentence "This is not a popular name of this year!" will be returned.
}
\description{
Enter the name and a particular year to see if the pet name is a popular name of that year.
The range of the year is from 2015 to 2018, as there are less than 20 observations
from the year prior 2015.
}
