% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet-name.R
\name{pet_summary}
\alias{pet_summary}
\title{The number summary of pet distribution in each year.}
\usage{
pet_summary(pet_count)
}
\arguments{
\item{pet_count}{is the name you want to check.}
}
\value{
The function will return a 5 columns data frame that shows year, mean, median,
standard deviation and the count of zipcode districts.
}
\description{
Enter a particular year to see mean, median and standard deviation of
each year. The range of the year is from 2015 to 2018, as there are
less than 20 observations from the year prior 2015.
}
