% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seattle_pets}
\alias{seattle_pets}
\title{Seattle pet name.}
\format{
A data frame with 52,122 observations and 7 variables:
\describe{
\item{license_issue_date}{The date when the licensse issued}
\item{license_number}{The number of that license}
\item{animals_name}{The name of the licensed pet}
\item{species}{The species of the licensed pet}
\item{primary_breed}{The primary breed of the licensed pet}
\item{secondary_breed}{The secondary breed of the licensed pet}
\item{zip_code}{The zipcode of the owner of the licensed pet}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-03-26}
\url{https://data.seattle.gov/Community/Seattle-Pet-Licenses/jguv-t9rb}
}
\usage{
seattle_pets
}
\description{
The data set contains the name of the licensed pet in Seattle.
}
\keyword{datasets}
