% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seattlepetname-package.R
\docType{package}
\name{seattlepetname-package}
\alias{seattlepetname}
\alias{seattlepetname-package}
\title{seattlepetname: The Distribution of Pets Across Seattle}
\description{
The package allows users to understand the distribution of pets across Seattle. Also, the user can understand the prefernece of pet owners for pet name and primary breeds in Seattle.
}
\author{
\strong{Maintainer}: Yan Chui Lucia Cheung \email{yche0164@student.monash.edu}

}
\keyword{internal}
