% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet-name.R
\name{top_name}
\alias{top_name}
\title{The top 10 popular pet name (by species) in Seattle.}
\usage{
top_name(pet_type)
}
\arguments{
\item{pet_type}{is the species of the pet.}
}
\value{
The function will return a interactive bar chart that
shows the count of the top 10 pet names.
}
\description{
Enter the species (Dog, cat, goat or pig) to see 10 most popular
pet names for that species.
}
